Transmit Event Continuous
  * Add code to transmit continuously between start and end time.
  * Add code to load the transmit queue with files on start-up and prior
    to issuing a timed event (user selectable).
  * Add code to manually load the transmit queue with queue list.
  * Add checkbox to load from '.flamp/tx' directory or use queue list.
  * Modified GUI Event Panel to support new Event.
  * Encapsulate rx_amp and tx_amp in global class.  Sync with mutex
    locks/unlocks.
  * On queue item removal refresh receive panel with newly assigned current
    queue item.
  * Added additional logging.
  * Fixed assignment issue with variable loading_from_queue_file
  * Removed FLMSG compress requirement.
  * Make a copy of path before passing to dirname().
  * Changed 'K' and 'BK' EOF logic for continuous transmit events
  * Redirect '<MISSING' tag detection from XML to ARQ data stream.
  * Fixed clear missing block(s) to operate on queued batch transfers.
  * Update TagSearch class to monitor for '<MISSING' tags.
  * Add function void process_missing_stream(void) into flamp.cxx to
    direct '<MISSING' tags to their respective cAmp instances.
  * Add function void cAmp::append_report(std::string s) into amp.cxx to
    accumulate missing reports.
  * Update ztimer() to call transmit_queued from the main thread using
    FL:awake(func). Keeping ztimer() in the one second execution window.
  * Add cAmp code to retain file timestamp in stat format.
  * Add cAmp code to check if file has been modified since last loaded.
  * Add two additional fields the the Queue List file; Compress and Description
  * Modify code in support of the new queue field items.
  * Correct midnight time transmit duration issue with 'continuos at' event.
  * Add visual indication for unproto use by deactivating selected widgets
    in the transmit panel.
  * New Event subpanel for selecting various modems for multimode multicasts.
  * Add code in support of multimodem multicast (Hamcast).
  * Fixed a bug in interval/header modem introduced in the last round of
    modifications.
  * Updated Modem selection to include MFSK64L and MFSK128L
  * Updated time table generator to support MFSK long modes.
  * Removed unused variables.
  * Move redundant code into common function.
  * New xmlrpc command 'main.get_tx_duration'. Requires FLDIGI update for
    timing table generation.
  * Add support directory and associated files.
  * Add 'rx/' to the save path (RX panel operation).
  * Saving event data from debug window produced an empty file, std::string
    dbg_buffer was never updated: Fixed. Added timestamp to output data.
  * Disable Events in Generate Time Table Mode.
  * Removed xmlrpc command 'main.get_tx_duration'
  * New xmlrpc command 'main.get_char_timing'
  * Add PSK250C6 mode to list.
  * Updated timing tables.
  * Correct typo in LOG_DEBUG message
  * Remove unused global string variable.
  * Update character 0x05 in MFSK timing tables.

+++ src/xmlrpcpp/XmlRpcClient.cpp
@@ -316,7 +316,7 @@ XmlRpcClient::generateHeader(std::string const& body)
 
   header += "Content-Type: text/xml\r\nContent-length: ";
 
-  sprintf(buff,"%d\r\n\r\n", body.size());
+  sprintf(buff,"%ld\r\n\r\n", (long) body.size());
 
   return header + buff;
 }
