+++ src/xmlrpcpp/XmlRpcServer.cpp
@@ -480,7 +480,7 @@ XmlRpcServer::generateHeader(std::string const& body)
     "Content-length: ";
 
   char buffLen[40];
-  sprintf(buffLen,"%d\r\n\r\n", body.size());
+  sprintf(buffLen,"%ld\r\n\r\n", (long) body.size());
 
   return header + buffLen;
 }
